refactor: As per upstream changes
