+++ .gitignore
@@ -3,3 +3,4 @@ frontend/bun.lockb
 *.db
 *.db.lock
 .DS_Store
+.venv
