+++ backend/.env.sample
@@ -1,25 +1,50 @@
 # -------------------------
 # LLM CONFIG
 # -------------------------
-DIMENSIONS=768
-EMBEDDING_PROVIDER=openai
+
 LLM_PROVIDER=openai
-OPENAI_API_KEY={your-openai-api-key}
 
 # -------------------------
 # VECTOR DATABASE CONFIG
 # -------------------------
+
 VECTOR_DB=milvus-lite
 INDEX_NAME={your-index-name}
+EMBEDDING_PROVIDER=openai
+DIMENSIONS=768
+
+# -------------------------
+# MILVUS CONFIG
+# Applicable if VECTOR_DB=milvus-lite
+# -------------------------
 MILVUS_DB_USERNAME=root
-MILVUS_DB_PASSWORD={your-milvus-password}
+
+# -------------------------
+# QDRANT CONFIG
+# Applicable if VECTOR_DB=qdrant
+# -------------------------
+
+# QDRANT_LOCATION="http://localhost:6333"
+# QDRANT_PORT=
+# QDRANT_GRPC_PORT=
+# QDRANT_PREFER_GRPC=
+# QDRANT_URL=
+# QDRANT_HTTPS=
+# QDRANT_API_KEY=
+# QDRANT_PREFIX=
+# QDRANT_TIMEOUT=
+# QDRANT_HOST=
+# QDRANT_PATH=
+
 
 # -------------------------
 # QUERY CONFIG
 # -------------------------
+
 QUERY_TYPE=hybrid
 
 # -------------------------
 # UNSTRUCTURED CONFIG
 # -------------------------
+
 UNSTRUCTURED_API_KEY={your-unstructured-api-key}
