+++ backend/.gitignore
@@ -165,5 +165,4 @@ cython_debug/
 #.idea/
 
 # Custom
-*milvus*
 .tests_static_analysis.sh
