+++ backend/src/knowledge_table_api/dependencies.py
@@ -4,6 +4,8 @@ from functools import lru_cache
 
 from pymilvus import MilvusClient
 
+from backend.src.knowledge_table_api.services.vector_index import VectorIndexFactory
+from backend.src.knowledge_table_api.services.vector_index.base import VectorIndex
 from knowledge_table_api.config import Settings
 from knowledge_table_api.services.llm_service import LLMFactory, LLMService
 
@@ -12,15 +14,6 @@ from knowledge_table_api.services.llm_service import LLMFactory, LLMService
 def get_settings() -> Settings:
     """Get the settings for the application."""
     return Settings()
-
-
-def get_milvus_client() -> MilvusClient:
-    """Get the Milvus client for the application."""
-    settings = get_settings()
-    return MilvusClient(
-        "./milvus_demo.db",
-        token=f"{settings.milvus_db_username}:{settings.milvus_db_password}",
-    )
 
 
 def get_llm_service() -> LLMService:
@@ -32,3 +25,12 @@ def get_llm_service() -> LLMService:
             f"Failed to create LLM service for provider: {settings.llm_provider}"
         )
     return llm_service
+
+
+def get_vector_index() -> VectorIndex:
+    """Get the vector index for the application."""
+    settings = get_settings()
+    vector_index = VectorIndexFactory.create_vector_index(settings.vector_db)
+    if vector_index is None:
+        raise ValueError(f"Invalid value for vector database : {settings.vector_db}")
+    return vector_index
