+++ backend/src/knowledge_table_api/services/vector_index/__init__.py
@@ -1,0 +1,19 @@
+import os
+from typing import Optional
+
+from .base import VectorIndex
+from .milvus import MilvusIndex
+from .qdrant import QdrantIndex
+
+
+class VectorIndexFactory:
+    """Factory for creating language model services."""
+
+    @staticmethod
+    def create_vector_index(provider: str = "milvus") -> Optional[VectorIndex]:
+        """Create a language model service."""
+        if provider == "milvus":
+            return MilvusIndex()
+        elif provider == "qdrant":
+            return QdrantIndex()
+        return None
